plugins/scribe/pyproject.toml
[project]
name = "scribe"
version = "1.7.3"
description = "Documentation review, cleanup, and generation with AI slop detection and style learning"
requires-python = ">=3.10"
dependencies = ["pyyaml"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".ruff_cache",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "mypy>=1.11.0",
    "ruff>=0.14.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --cov=scribe --cov-report=term-missing --cov-fail-under=90"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "bdd: Behavior-driven development style tests",
]

[tool.coverage.run]
source = ["src/scribe"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
]
fail_under = 90

[tool.mypy]
python_version = "3.9"

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["scribe.pattern_loader"]
disable_error_code = ["assignment"]

[[tool.mypy.overrides]]
module = ["scribe.tape_generator"]
disable_error_code = ["assignment"]
